feat: customer data from db and refactoring
