+++ frontend/src/pages/HomePage.tsx
@@ -1,8 +1,32 @@
 import {Button, Grid, Typography} from "@mui/material";
 import EnhancedTable from "../components/table/EnhancedTable";
-import {data} from "./data";
+import {useEffect, useState} from "react";
+import axios from "axios";
 
+export interface Customer {
+    id: string,
+    firstName: string,
+    lastName: string,
+    email: string,
+    phone: string,
+    street: string,
+    houseNo: number,
+    city: string,
+    postalCode: number,
+    linkToDSGVO: string,
+    createdBy: string,
+}
 export default function HomePage() {
+
+    const [customers, setCustomers] = useState<Customer[]>([]);
+
+    useEffect(() => {
+        (async () => {
+            const response = await axios.get("/api/customer")
+            setCustomers(response.data)
+
+        })();
+    }, [customers]);
 
     return (
         <Grid
@@ -17,7 +41,7 @@ export default function HomePage() {
 
             <Button sx={{my: 4}} variant="contained">Neuer Kunde</Button>
 
-            <EnhancedTable rows={data}/>
+            <EnhancedTable rows={customers}/>
         </Grid>
     )
 }
